refactor: improve code organization
